[fix] Don't fire multiple keydown events
Before jumping to the bug, let's thank `devcards` for making easy to find the
bug.

I noticed that pressing a key for the first time in `user-can-move-around` would
move the block one square, but the second time would move it two or three
squares.  Turned out that the callback for `keydown` was being registered more
than once.  `reagent` calls the function provided in `:ref` when the component
is mounted and when it is unmounted.  If the component was mounted but the node
didn't change, a new event was registered causing the "multiple" calls to
`on-change-direction`.  To fix this issue, I added one more atom to track the
last node seen and them register the events only when it is a new one.  **The
proper** fix is to create a wrapper that add and remove the listeners when the
component is mounted and unmounted (I will continue to research how to proper do
it).

After fixing the issue of "multiple key events", I discovered that the svg
wasn't been updated correctly and that is because a stateful reagent component
needs to return a render function and not the components.