[refactor] Remove `defevent!` macro
I have to structure the server in a way that I can re-use the event handling in
an "integration" test environment.  This macro is getting in the way to
accomplish it.  Besides that, `defmulti` and `defmethod` wide used in clojure
code and it will be easier for others to read these forms directly.